+++ types/mempool/mempool.go
@@ -7,10 +7,29 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
+type Tx struct {
+	Tx        sdk.Tx
+	GasWanted uint64
+}
+
+func NewMempoolTx(tx sdk.Tx, gasWanted uint64) Tx {
+	return Tx{
+		Tx:        tx,
+		GasWanted: gasWanted,
+	}
+}
+
+type GasTx interface {
+	GetGas() uint64
+}
+
 type Mempool interface {
 	// Insert attempts to insert a Tx into the app-side mempool returning
 	// an error upon failure.
 	Insert(context.Context, sdk.Tx) error
+
+	// Insert with a custom gas wanted value
+	InsertWithGasWanted(context.Context, sdk.Tx, uint64) error
 
 	// Select returns an Iterator over the app-side mempool. If txs are specified,
 	// then they shall be incorporated into the Iterator. The Iterator must
@@ -34,7 +53,7 @@ type Iterator interface {
 	Next() Iterator
 
 	// Tx returns the transaction at the current position of the iterator.
-	Tx() sdk.Tx
+	Tx() Tx
 }
 
 var (