+++ types/mempool/priority_nonce_test.go
@@ -386,7 +386,7 @@ func (s *MempoolTestSuite) TestIterator() {
 			// iterate through txs
 			iterator := pool.Select(ctx, nil)
 			for iterator != nil {
-				tx := iterator.Tx().(testTx)
+				tx := iterator.Tx().Tx.(testTx)
 				require.Equal(t, tt.txs[tx.id].p, int(tx.priority))
 				require.Equal(t, tt.txs[tx.id].n, int(tx.nonce))
 				require.Equal(t, tt.txs[tx.id].a, tx.address)
@@ -854,7 +854,7 @@ func TestNextSenderTx_TxReplacement(t *testing.T) {
 		require.Equal(t, 1, mp.CountTx())
 
 		iter := mp.Select(ctx, nil)
-		require.Equal(t, tx, iter.Tx())
+		require.Equal(t, tx, iter.Tx().Tx)
 	}
 
 	// test Priority with TxReplacement
@@ -889,5 +889,5 @@ func TestNextSenderTx_TxReplacement(t *testing.T) {
 	require.Equal(t, 1, mp.CountTx())
 
 	iter := mp.Select(ctx, nil)
-	require.Equal(t, txs[3], iter.Tx())
+	require.Equal(t, txs[3], iter.Tx().Tx)
 }